Moved hasAnId to CallbackValidator
